Separate fuelup self into command/ops
