+++ src/ops/mod.rs
@@ -1,1 +1,2 @@
+pub mod fuelup_self;
 pub mod fuelup_toolchain;
